rename package with logger implementation
